requirements.txt
beautifulsoup4==4.11.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
greenlet==1.1.3
idna==3.3
python-dotenv==0.21.0
requests==2.28.1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
urllib3==1.26.12
